Source code for Lighting system 2
This system has an INA219 current sensor interfaced with the NodeMCU. This sensor is used to calculate the battery percentage, power consumption, and power generation data. This data is pushed to the Firebase database. The credentials are hidden for security purposes. 
The INA219 sensor used I2C to communicate with the NodeMCU microcontroller. The SDA and SCL pins of the sensor are connected to the corresponding SDA and SCL pins on the microcontroller. When a user uses the mobile application to turn the LED ON/OFF or control the brightness that variable data is sent from the app to the Nodemcu through the firebase database. Those variables are set accordingly to control the LED.